% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcps2021.R
\name{gcps2021}
\alias{gcps2021}
\title{GCPS Report template for pdfs}
\usage{
gcps2021(...)
}
\arguments{
\item{...}{Other arguments to be passed to \code{base_format}.}
}
\description{
Some time in the future, this will help adding more features and templates
Right now it is not used
}
\details{
See `bookdown::pdf_book()` for options
}
